(#author) Implemented author with ChatGPT API
Got OpenAI API key, and stored it in `keys.py`. The file is not
uploaded in Github.

In `main.py`, added arguments for authors - model and temperature - and
implemented script-writing task using Author class.

Author class is implemented in `chatgpt.py`. It provides ChatGPT API an
instruction and an example of input and output script (which is in
`prompt.py`)

Also, added keyword to return value of crawler part.