(#editor) blurred edge, text on center, height limit
Removed background image.
Instead, resize blurred image to fill the screen.

Also, restricted minimum height of the images and videos, because too
wide video looks not quite good. Minimum height is equal to width of the
screen.

Finally, moved subtitles to the middle of the screen, as other shorts
videos.